% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/F0069.R
\name{add.10x.image}
\alias{add.10x.image}
\title{Add image data to iCellR object}
\usage{
add.10x.image(x = NULL, image.data.list = NULL, condition.names = NULL)
}
\arguments{
\item{x}{An object of class iCellR.}

\item{image.data.list}{A character vector of list object names. Lists should be made using "image.capture.10x" function. .}

\item{condition.names}{A character vector of condition names.}
}
\value{
An object of class iCellR
}
\description{
This function takes a list of image data adds it to the iCellR object.
}
