% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/F0068.R
\name{capture.image.10x}
\alias{capture.image.10x}
\title{Read 10X image data}
\usage{
capture.image.10x(dir.10x = NULL)
}
\arguments{
\item{dir.10x}{A directory that includes the 10X image files (scalefactors_json.json, tissue_lowres_image.png and tissue_positions_list.csv).}
}
\value{
A list object
}
\description{
This function takes 10X image data files and converts them to proper file format for iCellR.
}
