% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/F0045.R
\name{cc}
\alias{cc}
\title{Calculate Cell cycle phase prediction}
\usage{
cc(object = NULL, s.genes = s.phase, g2m.genes = g2m.phase)
}
\arguments{
\item{object}{A data frame containing gene counts for cells.}

\item{s.genes}{Genes that are used as a marker for S phase.}

\item{g2m.genes}{Genes that are used as a marker for G2 and M phase.}
}
\value{
The data frame object
}
\description{
This function takes an object of class iCellR and assignes cell cycle stage for the cells.
}
