% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/F0049.R
\name{cell.cycle}
\alias{cell.cycle}
\title{Cell cycle phase prediction}
\usage{
cell.cycle(
  object = NULL,
  scoring.List = NULL,
  return.stats = FALSE,
  scoring.method = "tirosh"
)
}
\arguments{
\item{object}{A data frame containing gene counts for cells.}

\item{scoring.List}{Genes that are used as a marker for phases.}

\item{return.stats}{Return the data or object. If FALSE the object would be returned.}

\item{scoring.method}{Choose from "coverage" or "tirosh" for scoring method. See: https://science.sciencemag.org/content/352/6282/189}
}
\value{
The data frame object
}
\description{
This function takes an object of class iCellR and assignes cell cycle stage for the cells.
}
