% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/F033.cell.type.pred.R
\name{cell.type.pred}
\alias{cell.type.pred}
\title{Create heatmaps or dot plots for genes in clusters to find thier cell types using ImmGen data.}
\usage{
cell.type.pred(immgen.data = "rna", gene = "NULL",
  top.cell.types = 50, plot.type = "heatmap", heat.colors = c("blue",
  "white", "red"))
}
\arguments{
\item{immgen.data}{Choose from "rna", "uli.rna" or "mca", default = "rna"}

\item{gene}{A set of gene names to used to predict cell type.}

\item{top.cell.types}{Top cell types sorted by cumulative expression, default = 25.}

\item{plot.type}{Choose from "heatmap" od "point.plot", default = "heatmap"}

\item{heat.colors}{Colors for heatmap, default = c("blue" ,"white", "red").}
}
\value{
An object of class iCellR
}
\description{
This function takes an object of class iCellR and genes and provides a heatmap.
}
\examples{
\dontrun{
imm.gen(immgen.data = "uli.rna", gene = MyGenes, plot.type = "heatmap")
imm.gen(immgen.data = "rna", gene = MyGenes, plot.type = "point.plot")
}
}
