% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/F0025.R
\name{change.clust}
\alias{change.clust}
\title{Change the cluster number or re-name them}
\usage{
change.clust(x = NULL, change.clust = 0, to.clust = 0, clust.reset = FALSE)
}
\arguments{
\item{x}{An object of class iCellR.}

\item{change.clust}{The name of the cluster to be changed.}

\item{to.clust}{The new name for the cluster.}

\item{clust.reset}{Reset to the original clustering.}
}
\value{
An object of class iCellR.
}
\description{
This function re-names the clusters in the best.clust slot of the iCellR object.
}
