% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/F021.clust.avg.exp.R
\name{clust.avg.exp}
\alias{clust.avg.exp}
\title{Create a data frame of mean expression of genes per cluster}
\usage{
clust.avg.exp(x = NULL)
}
\arguments{
\item{x}{An object of class iCellR.}
}
\value{
An object of class iCellR.
}
\description{
This function takes an object of class iCellR and creates an average gene expression for every cluster.
}
\examples{
\dontrun{
my.obj <- clust.avg.exp(my.obj)
}
}
