% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/F0050.R
\name{clust.ord}
\alias{clust.ord}
\title{Sort and relabel the clusters randomly or based on pseudotime}
\usage{
clust.ord(
  x = NULL,
  top.rank = 500,
  dist.method = "euclidean",
  clust.method = "complete",
  how.to.order = "distance"
)
}
\arguments{
\item{x}{An object of class iCellR.}

\item{top.rank}{A number. Taking the top genes ranked by base mean, default = 500.}

\item{dist.method}{Choose from "euclidean", "maximum", "manhattan", "canberra", "binary" or "minkowski", default = "euclidean".}

\item{clust.method}{Choose from "ward.D", "ward.D2", "single", "complete", "average", "mcquitty", "median" or "centroid", default = "complete".}

\item{how.to.order}{Choose from "distance" and "random".}
}
\value{
An object of class iCellR.
}
\description{
This function takes an object of class iCellR and re-ordersthe clusters based on pseudotime (distance).
}
