% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/F0026.R
\name{clust.rm}
\alias{clust.rm}
\title{Remove the cells that are in a cluster}
\usage{
clust.rm(x = NULL, clust.to.rm = "numeric")
}
\arguments{
\item{x}{A data frame containing gene counts for cells.}

\item{clust.to.rm}{The name of the cluster to be removed.}
}
\value{
An object of class iCellR
}
\description{
This function removes the cells from a designated cluster. Notice the cells will be removed from the main data (raw data would still have the original data).
}
