% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster.plot.R
\name{cluster.plot}
\alias{cluster.plot}
\title{Plot nGenes, UMIs and perecent mito}
\usage{
cluster.plot(x = NULL, cell.size = 1, plot.type = "tsne",
  cell.color = "black", back.col = "white", col.by = "clusters",
  cond.shape = F, cell.transparency = 0.5, clust.dim = 2,
  angle = 20, clonotype.max = 10, density = F, interactive = TRUE,
  static3D = F, out.name = "plot")
}
\arguments{
\item{x}{An object of class iCellR.}

\item{cell.size}{A numeric value for the size of the cells, defults = 1.}

\item{plot.type}{Choose between "tsne" and "pca", defult = "tsne".}

\item{cell.color}{Choose cell color if col.by = "monochrome", defult = "black".}

\item{back.col}{Choose background color, defult = "black".}

\item{col.by}{Choose between "clusters", "conditions" or "monochrome", defult = "clusters".}

\item{cell.transparency}{A numeric value between 0 to 1, defult = 0.5.}

\item{clust.dim}{A numeric value for plot dimentions. Choose either 2 or 3, defult = 2.}

\item{angle}{A number to rotate the non-interactive 3D plot.}

\item{density}{If TRUE the density plots for PCA/tSNE second dimention will be created, defult = FALSE.}

\item{interactive}{If TRUE an html intractive file will be made, defult = TRUE.}

\item{out.name}{Output name for html file if interactive = TRUE defult = "plot".}
}
\value{
An object of class iCellR.
}
\description{
This function takes an object of class iCellR and creats plot.
}
\examples{
\dontrun{
tsne.plot(my.obj)
}
}
