% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/F002.data.aggregation.R
\name{data.aggregation}
\alias{data.aggregation}
\title{Merge multiple data frames and add the condition names to their cell ids}
\usage{
data.aggregation(samples = NULL, condition.names = NULL)
}
\arguments{
\item{samples}{A character vector of data.frame object names.}

\item{condition.names}{A character vector of data.frame condition names.}
}
\value{
An object of class iCellR
}
\description{
This function takes data frame and merges them while also adding condition names to cell ids..
}
\examples{
\dontrun{
my.data <- data.aggregation(samples = c("sample1","sample2","sample3"),
                            condition.names = c("WT","KO","Ctrl"))
}

}
