% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/F0052.R
\docType{data}
\name{demo.obj}
\alias{demo.obj}
\title{An object of class iCellR for demo}
\format{
Subset of the data with 200 genes and 90 cells
}
\source{
\url{https://s3-us-west-2.amazonaws.com/10x.files/samples/cell/pbmc3k/pbmc3k_filtered_gene_bc_matrices.tar.gz}
}
\usage{
demo.obj
}
\description{
A demo object
}
\keyword{datasets}
