% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/F007.down.sample.R
\name{down.sample}
\alias{down.sample}
\title{Down sample conditions}
\usage{
down.sample(x = NULL)
}
\arguments{
\item{x}{An object of class iCellR.}
}
\value{
An object of class iCellR.
}
\description{
This function takes an object of class iCellR and down samples the condition to have equal number of cells in each condition.
}
\examples{
\dontrun{
my.obj <- down.sample(my.obj)
}
}
