% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find.dim.genes.R
\name{find.dim.genes}
\alias{find.dim.genes}
\title{Run PCA on the main data}
\usage{
find.dim.genes(x = NULL, dims = 1:10, top.pos = 15, top.neg = 5)
}
\arguments{
\item{x}{An object of class iCellR.}

\item{clust.method}{Choose from "base.mean.rank" or "gene.model", defult is "base.mean.rank".}

\item{top.rank}{A number taking the top genes ranked by base mean, defult = 500.}

\item{gene.list}{A list of genes to be used for PCA. If "clust.method" is set to "gene.model", defult = "my_model_genes.txt".}
}
\value{
An object of class iCellR.
}
\description{
This function takes an object of class iCellR and runs PCA on the main data.
}
\examples{
\dontrun{
my.obj <- run.pca(my.obj, clust.method = "gene.model", gene.list = "my_model_genes.txt")
}
}
