% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/F0053.R
\docType{data}
\name{g2m.phase}
\alias{g2m.phase}
\title{A dataset of G2 and M phase genes}
\format{
A character with 54 genes
}
\source{
\url{https://www.science.org/doi/abs/10.1126/science.aad0501}
}
\usage{
g2m.phase
}
\description{
A dataset containing the genes for G2 and M phase
}
\keyword{datasets}
