% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/F0071.R
\name{i.score}
\alias{i.score}
\title{Cell cycle phase prediction}
\usage{
i.score(
  object = NULL,
  data.type = "main.data",
  scoring.List = NULL,
  return.stats = TRUE,
  scoring.method = "tirosh"
)
}
\arguments{
\item{object}{A data frame containing gene counts for cells.}

\item{data.type}{Choose from "raw.data" or "main.data", "imputed.data", default = "main.data".}

\item{scoring.List}{Genes that are used as a marker for phases.}

\item{return.stats}{Return the data or object. If FALSE the object would be returned.}

\item{scoring.method}{Choose from "tirosh (Tirosh, et. al. 2016), mean, sum, gsva, ssgsea, zscore and plage. , default = "tirosh".}
}
\value{
The data frame object
}
\description{
This function takes an object of class iCellR and assignes cell cycle stage for the cells.
}
