% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/F0059.R
\name{iba}
\alias{iba}
\title{iCellR Batch Alignment (IBA)}
\usage{
iba(
  x = NULL,
  dims = 1:30,
  k = 10,
  ba.method = "CPCA",
  method = "base.mean.rank",
  top.rank = 500,
  plus.log.value = 0.1,
  scale.data = TRUE,
  gene.list = "character"
)
}
\arguments{
\item{x}{An object of class iCellR.}

\item{dims}{PC dimentions to be used}

\item{k}{number of neighboring cells for KNN, default = 10.}

\item{ba.method}{Batch alignment method. Choose from "CCCA" and "CPCA", default = "CPCA".}

\item{method}{Choose from "base.mean.rank" or "gene.model", default is "base.mean.rank". If gene.model is chosen you need to provide gene.list.}

\item{top.rank}{A number. Taking the top genes ranked by base mean, default = 500.}

\item{plus.log.value}{A number to add to each value in the matrix before log transformasion to aviond Inf numbers, default = 0.1.}

\item{scale.data}{If TRUE the data will be scaled (log2 + plus.log.value), default = TRUE.}

\item{gene.list}{A charactor vector of genes to be used for PCA. If "clust.method" is set to "gene.model", default = "my_model_genes.txt".}
}
\description{
This function takes an object of class iCellR and runs CCCA or CPCA batch alignment.
}
