% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/F0064.R
\name{load.h5}
\alias{load.h5}
\title{Load h5 data as data.frame}
\usage{
load.h5(filename, feature.names = TRUE, uniq.rows = TRUE)
}
\arguments{
\item{filename}{path to the input (h5) file}

\item{feature.names}{row names to be feature names or ID numbers.}

\item{uniq.rows}{make row names unique.}
}
\value{
The data frame object
}
\description{
This function reads hdf5 files.
}
