% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/F001.load10x.R, R/F1.load10x.R
\name{load10x}
\alias{load10x}
\title{Load 10X data as data.frame}
\usage{
load10x(dir.10x = NULL, gene.name = "geneSymbol")

load10x(dir.10x = NULL, gene.name = "geneSymbol")
}
\arguments{
\item{dir.10x}{A directory that includes the 10X barcodes.tsv, genes.tsv and matrix.mtx files.}

\item{gene.name}{Should be either geneSymbol or ensembleID.}

\item{dir.10x}{A directory that includes the 10X barcodes.tsv, genes.tsv and matrix.mtx files.}

\item{gene.name}{Should be either geneSymbol or ensembleID.}
}
\value{
The data frame object

The data frame object
}
\description{
This function takes 10X data files barcodes.tsv, genes.tsv and matrix.mtx and converts them to proper matrix file for iCellR.

This function takes 10X data files barcodes.tsv, genes.tsv and matrix.mtx and converts them to proper matrix file for iCellR.
}
\examples{
\dontrun{
load10x("/hg19", gene.name = "geneSymbol")
}
\dontrun{
load10x("/hg19", gene.name = "geneSymbol")
}
}
