% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lvbsc.R
\name{lvbsc}
\alias{lvbsc}
\title{Low Variance Batch Spacing Correction (LVBSC)}
\usage{
lvbsc(x = NULL, dispersion.limit.max = 1.5, base.mean.rank = 500)
}
\arguments{
\item{x}{An object of class iCellR.}

\item{dispersion.limit.max}{A number for taking the genes that have dispersion below this number, defult = 1.5.}

\item{base.mean.rank}{A number taking the top genes ranked by base mean, defult = 500.}

\item{non.sig.col}{Color for the genes not used for the model, defult = "darkgray".}

\item{right.sig.col}{Color for the genes above the dispersion limit, defult = "chartreuse3".}

\item{left.sig.col}{Color for the genes above the rank limit, defult = "cadetblue3".}

\item{disp.line.col}{Color of the line for dispersion limit, defult = "black".}

\item{rank.line.col}{Color of the line for rank limit, defult = "red".}

\item{cell.size}{A number for the size of the points in the plot, defult = 1.75.}

\item{no.mito.model}{If set to TRUE, mitochondrial genes would be excluded from the gene list made for clustering, defult = TRUE.}

\item{mark.mito}{Mark mitochondrial genes in the plot, defult = TRUE.}

\item{cell.transparency}{Color transparency for the points in the plot, defult = 0.5.}

\item{interactive}{If set to TRUE an intractive HTML file will be created, defult = TRUE.}

\item{out.name}{If "interactive" is set to TRUE, the out put name for HTML, defult = "plot".}
}
\value{
An object of class iCellR.
}
\description{
This function takes an object of class iCellR and finds the top genes with low variance across the samples to make a model for batch spacing correction.
}
\examples{
\dontrun{
lvbsc(my.obj,
               dispersion.limit.max = 1.5,
               base.mean.rank = 500,
               no.mito.model = T,
               mark.mito = T,
               interactive = T,
               out.name = "gene.model")

lvbsc(my.obj,
             dispersion.limit.max = 1.5,
             base.mean.rank = 500)
}

}
