% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/F0052.R
\name{make.bed}
\alias{make.bed}
\title{Make BED Files}
\usage{
make.bed(x = NULL)
}
\arguments{
\item{x}{Peak marker file.}
}
\value{
Bed files
}
\description{
This function takes peak marker files and makes the bed files per cluster.
}
