% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/man.assign.clust.R
\name{man.assign.clust}
\alias{man.assign.clust}
\title{Hierarchical clustering based on tSNE}
\usage{
man.assign.clust(x = NULL, clust.num = 0)
}
\arguments{
\item{x}{An object of class iCellR.}

\item{clust.num}{Number of clusters to be made.}
}
\value{
An object of class iCellR.
}
\description{
This function takes an object of class iCellR and performs hierarchical clustering based on tSNE data.
}
\examples{
\dontrun{
assign.clust(my.obj, clust.num = 7)
}
}
