% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myImp.R
\name{myImp}
\alias{myImp}
\title{Merge RNA and ADT data}
\usage{
myImp(x = NULL)
}
\arguments{
\item{x}{An object of class iCellR.}

\item{adt.data}{Choose from raw or main (normalized) ADT data, default = "raw".}
}
\value{
An object of class iCellR
}
\description{
This function is to merge the RNA and ADT data to the main.data slot of the iCellR object.
}
\examples{
\dontrun{
my.obj <- myImp(my.obj)
}

}
