% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/F037.norm.adt.R
\name{norm.adt}
\alias{norm.adt}
\title{Normalize ADT data.
This function takes data frame and Normalizes ADT data.}
\usage{
norm.adt(x = NULL)
}
\arguments{
\item{x}{An object of class iCellR.}
}
\value{
An object of class iCellR
}
\description{
Normalize ADT data.
This function takes data frame and Normalizes ADT data.
}
\examples{
\dontrun{
my.obj <- norm.adt(my.obj)
}

}
