% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc.stats.R
\name{qc.stats}
\alias{qc.stats}
\title{Calculate the number of UMIs and genes per cell and percentage of mitochondrial genes per cell.}
\usage{
qc.stats(x = NULL, which.data = "raw.data", mito.genes = "defult",
  s.phase.genes = s.phase, g2m.phase.genes = g2m.phase)
}
\arguments{
\item{x}{A data frame containing gene counts for cells.}
}
\value{
The data frame object
}
\description{
This function takes data frame and calculates the number of UMIs, genes per cell and percentage of mitochondrial genes per cell.
}
\examples{
\dontrun{
UMIs.genes.mit(my.data)
}
}
