% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.anchor.R
\name{run.anchor}
\alias{run.anchor}
\title{Run anchor alignment on the main data.}
\usage{
run.anchor(x = NULL, method = "base.mean.rank", top.rank = 500,
  gene.list = "character", data.type = "main",
  normalization.method = "LogNormalize", scale.factor = 10000,
  margin = 1, block.size = NULL, selection.method = "vst",
  nfeatures = 2000, anchor.features = 2000, scale = TRUE,
  sct.clip.range = NULL, reduction = c("cca", "rpca"),
  l2.norm = TRUE, dims = 1:30, k.anchor = 5, k.filter = 200,
  k.score = 30, max.features = 200, nn.method = "rann", eps = 0,
  k.weight = 100, verbose = TRUE)
}
\arguments{
\item{x}{An object of class iCellR.}

\item{method}{Choose from "base.mean.rank" or "gene.model", default is "base.mean.rank". If gene.model is chosen you need to provide gene.list.}

\item{top.rank}{A number taking the top genes ranked by base mean, default = 500.}

\item{gene.list}{A charactor vector of genes to be used for PCA. If "clust.method" is set to "gene.model", default = "my_model_genes.txt".}

\item{data.type}{Choose from "main" and "imputed", default = "main"}
}
\value{
An object of class iCellR.
}
\description{
This function takes an object of class iCellR and runs anchor alignment. It's a wrapper for Seurat.
}
