% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.cca.R
\name{run.cca}
\alias{run.cca}
\title{Run PCA on the main data}
\usage{
run.cca(x = NULL, top.vari.genes = 1000, cc.number = 30,
  dims.align = 1:20, normalize.data = TRUE, scale.data = TRUE,
  normalization.method = "LogNormalize", scale.factor = 10000,
  display.progress = TRUE)
}
\arguments{
\item{x}{An object of class iCellR.}

\item{clust.method}{Choose from "base.mean.rank" or "gene.model", default is "base.mean.rank".}

\item{top.rank}{A number taking the top genes ranked by base mean, default = 500.}

\item{plus.log.value}{A number to add to each value in the matrix before log transformasion to aviond Inf numbers, default = 0.1.}

\item{gene.list}{A list of genes to be used for PCA. If "clust.method" is set to "gene.model", default = "my_model_genes.txt".}
}
\value{
An object of class iCellR.
}
\description{
This function takes an object of class iCellR and runs PCA on the main data.
}
\examples{
\dontrun{
my.obj <- run.pca(my.obj, clust.method = "gene.model", gene.list = "my_model_genes.txt")
}
}
