% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.diffusion.map.R
\name{run.diffusion.map}
\alias{run.diffusion.map}
\title{Run PHATE on PCA data (PHATE - Potential of Heat-Diffusion for Affinity-Based Transition Embedding)}
\usage{
run.diffusion.map(x = NULL, dims = 1:10, method = "phate",
  ndim = 3, k = 5, alpha = 40, n.landmark = 2000, gamma = 1,
  t = "auto", knn.dist.method = "euclidean", init = NULL,
  mds.method = "metric", mds.dist.method = "euclidean", t.max = 100,
  npca = 100, plot.optimal.t = FALSE, verbose = 1, n.jobs = 1,
  seed = NULL, potential.method = NULL, use.alpha = NULL,
  n.svd = NULL, pca.method = NULL, g.kernel = NULL, diff.op = NULL,
  landmark.transitions = NULL, diff.op.t = NULL, dist.method = NULL)
}
\arguments{
\item{x}{An object of class iCellR.}

\item{dims}{PC dimentions to be used for UMAP analysis.}
}
\value{
An object of class iCellR.
}
\description{
This function takes an object of class iCellR and runs PHATE on PCA data.
}
\examples{
\dontrun{
my.obj <- run.diffusion.map(my.obj, dims = 1:10, method = "phate")
}
}
