% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.magic.R
\name{run.magic}
\alias{run.magic}
\title{Run UMAP on PCA data}
\usage{
run.magic(x = NULL)
}
\arguments{
\item{x}{An object of class iCellR.}

\item{dims}{PC dimentions to be used for UMAP analysis.}
}
\value{
An object of class iCellR.
}
\description{
This function takes an object of class iCellR and runs UMAP on PCA data.
}
\examples{
\dontrun{
my.obj <- run.magic(my.obj)
}
}
