% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/F012.run.pca.R
\name{run.pca}
\alias{run.pca}
\title{Run PCA on the main data}
\usage{
run.pca(x = NULL, data.type = "main", method = "base.mean.rank",
  top.rank = 500, plus.log.value = 0.1, batch.norm = F,
  gene.list = "character")
}
\arguments{
\item{x}{An object of class iCellR.}

\item{method}{Choose from "base.mean.rank" or "gene.model", default is "base.mean.rank". If gene.model is chosen you need to provide gene.list.}

\item{top.rank}{A number taking the top genes ranked by base mean, default = 500.}

\item{plus.log.value}{A number to add to each value in the matrix before log transformasion to aviond Inf numbers, default = 0.1.}

\item{batch.norm}{If TRUE the data will be normalized based on the genes in gene.list or top ranked genes.}

\item{gene.list}{A charactor vector of genes to be used for PCA. If "clust.method" is set to "gene.model", default = "my_model_genes.txt".}
}
\value{
An object of class iCellR.
}
\description{
This function takes an object of class iCellR and runs PCA on the main data.
}
\examples{
\dontrun{
my.obj <- run.pca(my.obj, clust.method = "gene.model", gene.list = "my_model_genes.txt")
}
}
