% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/F0067.R
\name{run.phenograph}
\alias{run.phenograph}
\title{Clustering the data}
\usage{
run.phenograph(x = NULL, k = 100, data.type = "pca", dims = 1:10)
}
\arguments{
\item{x}{An object of class iCellR.}

\item{k}{integer; number of nearest neighbours (default:45)}

\item{data.type}{Choose between "tsne", "pca", "umap", default = "pca".}

\item{dims}{PCA dimentions to be use for clustering, default = 1:10.}
}
\value{
An object of class iCellR.
}
\description{
This function takes an object of class iCellR and finds optimal number of clusters and clusters the data.
}
