% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/F029.top.markers.R
\name{top.markers}
\alias{top.markers}
\title{Choose top marker genes}
\usage{
top.markers(x = NULL, topde = 10, min.base.mean = 0.2, cluster = 0)
}
\arguments{
\item{x}{An object of class iCellR.}

\item{topde}{Number of top differentially expressed genes to be choosen from each cluster, default = 10.}

\item{min.base.mean}{Minimum base mean of the genes to be chosen, default = 0.5.}
}
\value{
A set of gene names
}
\description{
This function takes the marker genes info if chooses marker gene names for plots.
}
\examples{
\dontrun{
MyGenes <- top.markers(marker.genes, topde = 10, min.base.mean = 0.8)
}
}
