% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/F042vdj.stats.R
\name{vdj.stats}
\alias{vdj.stats}
\title{VDJ stats}
\usage{
vdj.stats(vdj.data = "data.frame")
}
\arguments{
\item{vdj.data}{A data frame containing VDJ data for cells.}
}
\value{
An object of class iCellR
}
\description{
This function takes a data frame of VDJ info per cell and dose QC.
}
\examples{
\dontrun{
Read your VDJ data (in this case in VDJ.tsv file) and add to your object as below

my.vdj.data <- read.table("VDJ.tsv")

VDJ <- prep.vdj(my.obj, adt.data = my.vdj.data)

head(VDJ)

vdj.stats(vdj.data = VDJ)
}

}
