% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/F040.add.vdj.R
\name{add.vdj}
\alias{add.vdj}
\title{Add V(D)J recombination data}
\usage{
add.vdj(x = NULL, vdj.data = "data.frame")
}
\arguments{
\item{x}{An object of class iCellR.}

\item{adt.data}{A data frame containing VDJ information for cells.}
}
\value{
An object of class iCellR
}
\description{
This function takes a data frame of VDJ information per cell and adds it to the iCellR object.
}
\examples{
\dontrun{
my.obj <- add.adt(my.obj, adt.data = adt.data)
}

}
