% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/F0023.R
\name{cell.gating}
\alias{cell.gating}
\title{Cell gating}
\usage{
cell.gating(x = NULL, my.plot = NULL, plot.type = "tsne")
}
\arguments{
\item{x}{An object of class iCellR.}

\item{my.plot}{The plot to use for gating. Must be a 2D plot.}

\item{plot.type}{Choose from knetl, umap and tsne, default = NULL.}
}
\value{
An object of class iCellR.
}
\description{
This function takes an object of class iCellR and a 2D tSNE or UMAP plot and gates around cells to get their ids.
}
