% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clono.plot.R
\name{clono.plot}
\alias{clono.plot}
\title{Make scatter, box and bar plots for genes}
\usage{
clono.plot(x = NULL, plot.data.type = "tsne", clono = 1,
  clust.dim = 2, cell.size = 1, cell.colors = c("red", "gray"),
  box.cell.col = "black", back.col = "white",
  cell.transparency = 0.5, interactive = TRUE, out.name = "plot")
}
\arguments{
\item{x}{An object of class scSeqR.}

\item{plot.data.type}{Choose from "tsne" and "pca", defult = "tsne".}

\item{clust.dim}{2 for 2D plots and 3 for 3D plots, defult = 2.}

\item{cell.size}{A number for the size of the points in the plot, defult = 1.}

\item{cell.colors}{Colors for heat mapping the points in "scatterplot", defult = c("gray","red").}

\item{box.cell.col}{A color for the points in the box plot, defult = "black".}

\item{back.col}{A color for the plot background, defult = "black".}

\item{cell.transparency}{Color transparency for points in "scatterplot" and "boxplot", defult = 0.5.}

\item{interactive}{If set to TRUE an intractive HTML file will be created, defult = TRUE.}

\item{out.name}{If "interactive" is set to TRUE, the out put name for HTML, defult = "plot".}

\item{gene}{A gene name to be plotted.}

\item{box.to.test}{A cluster number so that all the boxes in the box plot would be compared to. If set to "0" the cluster with the highest avrage would be choosen, defult = 0.}

\item{box.pval}{Choose from "sig.values" and "sig.signs". If set to "sig.signs" p values would be replaced with signs ("na", "*", "**", "***"), defult = "sig.signs".}

\item{col.by}{Choose from "clusters" and "conditions", defult = "clusters".}

\item{plot.type}{Choose from "scatterplot", "boxplot" and "barplot", defult = "scatterplot".}

\item{box.color}{A color for the boxes in the "boxplot", defult = "red".}

\item{box.line.col}{A color for the lines around the "boxplot", defult = "green".}
}
\value{
An object of class scSeqR.
}
\description{
This function takes an object of class scSeqR and provides plots for genes.
}
\examples{
\dontrun{
clono.plot(my.obj,
            plot.data.type = "tsne",
            clono = 1,
            clust.dim = 2,
            cell.size = 1,
            cell.colors = c("red","gray"),
            cbox.cell.col = "black",
            back.col = "white",
            interactive = T,
            cell.transparency = 0.5,
            out.name = "tSNE_3D_clusters")
}
}
