% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/F0021.R
\name{clust.avg.exp}
\alias{clust.avg.exp}
\title{Create a data frame of mean expression of genes per cluster}
\usage{
clust.avg.exp(x = NULL, data.type = "main", conds.to.avg = NULL)
}
\arguments{
\item{x}{An object of class iCellR.}

\item{data.type}{Choose from "main" and "imputed", default = "main"}

\item{conds.to.avg}{Choose the conditions you want to average, default = NULL (all conditions).}
}
\value{
An object of class iCellR.
}
\description{
This function takes an object of class iCellR and creates an average gene expression for every cluster.
}
\examples{
demo.obj <- clust.avg.exp(demo.obj)

head(demo.obj@clust.avg)
}
