% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/F009.R
\name{data.scale}
\alias{data.scale}
\title{Scale data}
\usage{
data.scale(x = NULL)
}
\arguments{
\item{x}{An object of class iCellR.}
}
\value{
An object of class iCellR.
}
\description{
This function takes an object of class iCellR and scales the normalized data.
}
\examples{
my.obj <- data.scale(demo.obj)

head(my.obj@scaled.data)[1:5]

}
