% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/F0014.R
\name{find.dim.genes}
\alias{find.dim.genes}
\title{Find model genes from PCA data}
\usage{
find.dim.genes(x = NULL, dims = 1:10, top.pos = 15, top.neg = 5)
}
\arguments{
\item{x}{An object of class iCellR.}

\item{dims}{PC dimentions to be used.}

\item{top.pos}{Number of top positive marker genes to be taken from each PC, default = 15.}

\item{top.neg}{Number of top negative marker genes to be taken from each PC, default = 5.}
}
\value{
An object of class iCellR.
}
\description{
This function takes an object of class iCellR finds the model genes to run a second round of PCA.
}
\examples{

demo.obj <- find.dim.genes(demo.obj, dims = 1:10,top.pos = 20, top.neg = 20)

head(demo.obj@gene.model)

}
