% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findMarkers.R
\name{findMarkers}
\alias{findMarkers}
\title{Find marker genes for each cluster}
\usage{
findMarkers(x = NULL, fold.change = 2, padjval = 0.1, Inf.FCs = F,
  uniq = F, positive = TRUE)
}
\arguments{
\item{x}{An object of class iCellR.}

\item{fold.change}{A number that designates the minimum fold change for out put, defult = 2.}

\item{padjval}{Minimum adjusted p value for out put, defult = 0.1.}

\item{Inf.FCs}{If set to FALSE the infinate fold changes would be filtered from out put, defult = FALSE.}

\item{uniq}{If set to TRUE only genes that are a marker for only one cluster would be in the out put, defult = TRUE.}

\item{positive}{If set to FALSE both the up regulated (positive) and down regulated (negetive) markers would be in the out put, defult = FALSE.}
}
\value{
An object of class iCellR
}
\description{
This function takes an object of class iCellR and performs differential expression (DE) analysis to find marker genes for each cluster.
}
\examples{
\dontrun{
marker.genes <- find.markers(my.obj,fold.change = 2,padjval = 0.1,uniq = T)
}

}
