% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/F0053.R
\docType{data}
\name{g2m.phase}
\alias{g2m.phase}
\title{A dataset of G2 and M phase genes}
\format{
A character with 54 genes
}
\source{
Paper: Dissecting the multicellular ecosystem of metastatic melanoma by single-cell RNA-seq.
}
\usage{
g2m.phase
}
\description{
A dataset containing the genes for G2 and M phase
}
\keyword{datasets}
