% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/F031.gene.plot.R
\name{gene.plot}
\alias{gene.plot}
\title{Make scatter, box and bar plots for genes}
\usage{
gene.plot(x = NULL, gene = "NULL", data.type = "main",
  box.to.test = 0, box.pval = "sig.signs", plot.data.type = "tsne",
  clust.dim = 2, col.by = "clusters", plot.type = "scatterplot",
  cell.size = 1, cell.colors = c("gray", "red"),
  box.cell.col = "black", box.color = "red", box.line.col = "green",
  back.col = "white", cell.transparency = 0.5, interactive = TRUE,
  out.name = "plot")
}
\arguments{
\item{x}{An object of class iCellR.}

\item{gene}{A gene name to be plotted.}

\item{box.to.test}{A cluster number so that all the boxes in the box plot would be compared to. If set to "0" the cluster with the highest avrage would be choosen, default = 0.}

\item{box.pval}{Choose from "sig.values" and "sig.signs". If set to "sig.signs" p values would be replaced with signs ("na", "*", "**", "***"), default = "sig.signs".}

\item{plot.data.type}{Choose from "tsne" and "pca", default = "tsne".}

\item{clust.dim}{2 for 2D plots and 3 for 3D plots, default = 2.}

\item{col.by}{Choose from "clusters" and "conditions", default = "clusters".}

\item{plot.type}{Choose from "scatterplot", "boxplot" and "barplot", default = "scatterplot".}

\item{cell.size}{A number for the size of the points in the plot, default = 1.}

\item{cell.colors}{Colors for heat mapping the points in "scatterplot", default = c("gray","red").}

\item{box.cell.col}{A color for the points in the box plot, default = "black".}

\item{box.color}{A color for the boxes in the "boxplot", default = "red".}

\item{box.line.col}{A color for the lines around the "boxplot", default = "green".}

\item{back.col}{A color for the plot background, default = "black".}

\item{cell.transparency}{Color transparency for points in "scatterplot" and "boxplot", default = 0.5.}

\item{interactive}{If set to TRUE an interactive HTML file will be created, default = TRUE.}

\item{out.name}{If "interactive" is set to TRUE, the out put name for HTML, default = "plot".}
}
\value{
An object of class iCellR.
}
\description{
This function takes an object of class iCellR and provides plots for genes.
}
\examples{
\dontrun{
cluster.plot(my.obj,
            cell.size = 1,
            plot.type = "tsne",
            cell.color = "black",
            back.col = "white",
            col.by = "clusters",
            cell.transparency = 0.5,
            clust.dim = 3,
            interactive = T,
            density = F,
            out.name = "tSNE_3D_clusters")

cluster.plot(my.obj, cell.size = 1,
            plot.type = "tsne",
            col.by = "clusters",
            cell.transparency = 0.5,
            clust.dim = 2,
            interactive = T,
            density = F,
            out.name = "tSNE_2D_clusters")

cluster.plot(my.obj,
           cell.size = 2,
           plot.type = "tsne",
           clust.dim = 2,
           interactive = F)

cluster.plot(my.obj,
           cell.size = 1,
           plot.type = "tsne",
           col.by = "clusters",
           clust.dim = 3,
           interactive = F,
           angle = 45)


cluster.plot(my.obj,
          cell.size = 1,
          plot.type = "pca",
          cell.color = "black",
          back.col = "white",
          col.by = "conditions",
          cell.transparency = 0.5,
          clust.dim = 3,
          interactive = T,
          density = F,
          out.name = "PCA_3D_conditions")
}
}
