% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/F0010.R
\name{gene.stats}
\alias{gene.stats}
\title{Make statistical information for each gene across all the cells (SD, mean, expression, etc.)}
\usage{
gene.stats(x = NULL, which.data = "raw.data", each.cond = FALSE)
}
\arguments{
\item{x}{An object of class iCellR.}

\item{which.data}{Choose from "raw.data" or "main.data", default = "raw.data".}

\item{each.cond}{If TRUE each condition will be calculated, default = FALSE.}
}
\value{
An object of class iCellR.
}
\description{
This function takes an object of class iCellR and provides some statistical information for the genes.
}
\examples{
demo.obj <- gene.stats(demo.obj, which.data = "main.data")
head(demo.obj@gene.data)

}
