% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/F0055.R
\name{gg.cor}
\alias{gg.cor}
\title{Gene-gene correlation.
This function helps to visulaize and calculate gene-gene correlations.}
\usage{
gg.cor(
  x = NULL,
  data.type = "imputed",
  gene1 = NULL,
  gene2 = NULL,
  conds = NULL,
  clusts = NULL,
  cell.size = 1,
  cell.transparency = 0.5,
  interactive = TRUE,
  out.name = "plot"
)
}
\arguments{
\item{x}{An object of class iCellR.}

\item{data.type}{Choose from imputed and main, default = "imputed".}

\item{gene1}{First gene name.}

\item{gene2}{Second gene name.}

\item{conds}{Filter only one condition (only one), default is all conditions.}

\item{clusts}{Choose clusters to plot.}

\item{cell.size}{A numeric value for the size of the cells, default = 1.}

\item{cell.transparency}{A numeric value between 0 to 1, default = 0.5.}

\item{interactive}{If TRUE an html interactive file will be made, default = TRUE.}

\item{out.name}{Output name for html file if interactive = TRUE, default = "plot".}
}
\value{
An object of class iCellR
}
\description{
Gene-gene correlation.
This function helps to visulaize and calculate gene-gene correlations.
}
