% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmap.gg.plot.R
\name{heatmap.gg.plot}
\alias{heatmap.gg.plot}
\title{Create heatmaps for genes in clusters or conditions.}
\usage{
heatmap.gg.plot(x = NULL, gene = "NULL", cluster.by = "clusters",
  min.scale = -2.5, max.scale = 2.5, interactive = T, cex.col = 10,
  cex.row = 10, no.key = FALSE, out.name = "plot",
  heat.colors = c("blue", "white", "red"))
}
\arguments{
\item{x}{A data frame containing gene counts for cells.}

\item{gene}{A set of gene names to be heatmapped.}

\item{cluster.by}{Choose from "clusters" or "conditions", defult = "clusters".}

\item{interactive}{If TRUE an html intractive file will be made, defult = TRUE.}

\item{no.key}{If you want a color legend key defult = FALSE.}

\item{out.name}{Output name for html file if interactive = TRUE defult = "plot".}

\item{heat.colors}{Colors for heatmap, defult = c("blue" ,"white", "red").}
}
\value{
An object of class iCellR
}
\description{
This function takes an object of class iCellR and genes and provides a heatmap.
}
\examples{
\dontrun{
heatmap.gg.plot(my.obj, gene = MyGenes, interactive = T, out.name = "plot", cluster.by = "clusters")
}
}
