% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/F0013.R
\name{opt.pcs.plot}
\alias{opt.pcs.plot}
\title{Find optimal number of PCs for clustering}
\usage{
opt.pcs.plot(x = NULL, pcs.in.plot = 50)
}
\arguments{
\item{x}{An object of class iCellR.}

\item{pcs.in.plot}{Number of PCs to show in plot, defult = 50.}
}
\value{
An object of class iCellR.
}
\description{
This function takes an object of class iCellR and finds optimal number of PCs for clustering.
}
\examples{
opt.pcs.plot(demo.obj)

}
