% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/F0039.R
\name{prep.vdj}
\alias{prep.vdj}
\title{Prepare VDJ data}
\usage{
prep.vdj(vdj.data = "data.frame", cond.name = "NULL")
}
\arguments{
\item{vdj.data}{A data frame containing vdj information.}

\item{cond.name}{Conditions.}
}
\value{
An object of class iCellR
}
\description{
This function takes a data frame of VDJ data per cell and prepares it to adds it to the iCellR object.
}
