% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/F039.prep.vdj.R
\name{prep.vdj}
\alias{prep.vdj}
\title{Add CITE-seq antibody-derived tags (ADT)}
\usage{
prep.vdj(vdj.data = "all_contig_annotations.csv", cond.name = "NULL")
}
\arguments{
\item{x}{An object of class iCellR.}

\item{adt.data}{A data frame containing ADT counts for cells.}
}
\value{
An object of class iCellR
}
\description{
This function takes a data frame of ADT values per cell and adds it to the iCellR object.
}
\examples{
\dontrun{
my.obj <- add.adt(my.obj, adt.data = adt.data)
}

}
