% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/F0062.R
\name{pseudotime}
\alias{pseudotime}
\title{Pseudotime}
\usage{
pseudotime(x = NULL, marker.genes = "NULL", dims = 1:10)
}
\arguments{
\item{x}{An object of class iCellR.}

\item{marker.genes}{A list of marker genes for clusters.}

\item{dims}{PC dimentions to be used, , default = 1:10.}
}
\value{
An object of class iCellR.
}
\description{
This function takes an object of class iCellR and marker genes for clusters and performs pseudotime analysis.
}
