% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/F004.R
\name{qc.stats}
\alias{qc.stats}
\title{Calculate the number of UMIs and genes per cell and percentage of mitochondrial genes per cell and cell cycle genes.}
\usage{
qc.stats(
  x = NULL,
  which.data = "raw.data",
  mito.genes = NULL,
  s.phase.genes = s.phase,
  g2m.phase.genes = g2m.phase
)
}
\arguments{
\item{x}{A data frame containing gene counts for cells.}

\item{which.data}{Choose from raw data or main data, default = "raw.data".}

\item{mito.genes}{A character vector of mitochondrial  genes names , default is the genes starting with mt.}

\item{s.phase.genes}{A character vector of gene names for S phase, default = s.phase.}

\item{g2m.phase.genes}{A character vector of gene names for G2 and M phase, default = g2m.phase.}
}
\value{
The data frame object
}
\description{
This function takes data frame and calculates the number of UMIs, genes per cell and percentage of mitochondrial genes per cell and cell cycle genes.
}
\examples{
New.demo.obj <- qc.stats(demo.obj)
head(New.demo.obj@stats)
}
