% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/F0063.R
\name{run.cca}
\alias{run.cca}
\title{Run CCA on the main data}
\usage{
run.cca(
  x = NULL,
  top.vari.genes = 1000,
  cc.number = 30,
  dims.align = 1:20,
  normalize.data = TRUE,
  scale.data = TRUE,
  normalization.method = "LogNormalize",
  scale.factor = 10000,
  display.progress = TRUE
)
}
\arguments{
\item{x}{An object of class iCellR.}

\item{top.vari.genes}{Chose top genes to use for CCA, default = 1000.}

\item{cc.number}{Choose a number, default = 30.}

\item{dims.align}{Choose the CCA dimentions to align, default = 1:20.}

\item{normalize.data}{TRUE or FALSE, default = TRUE.}

\item{scale.data}{TRUE or FALSE, default = TRUE.}

\item{normalization.method}{Choose a method, default = "LogNormalize".}

\item{scale.factor}{Scaling factor, default = 10000.}

\item{display.progress}{Show progress, default = TRUE.}
}
\value{
An object of class iCellR.
}
\description{
This function takes an object of class iCellR and runs CCA using Seurat.
}
