% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/F0051.R
\docType{data}
\name{s.phase}
\alias{s.phase}
\title{A dataset of S phase genes}
\format{
A character with 43 genes
}
\source{
\url{https://www.science.org/doi/abs/10.1126/science.aad0501}
}
\usage{
s.phase
}
\description{
A dataset containing the genes for S phase
}
\keyword{datasets}
