% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/F0042.R
\name{vdj.stats}
\alias{vdj.stats}
\title{VDJ stats}
\usage{
vdj.stats(my.vdj = "data.frame")
}
\arguments{
\item{my.vdj}{A data frame containing VDJ data for cells.}
}
\value{
An object of class iCellR
}
\description{
This function takes a data frame of VDJ info per cell and dose QC.
}
